requirements.txt
openai==1.34.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.7
langchain-openai==0.1.8
streamlit==1.35.0
huggingface_hub==0.23.4
python-dotenv==1.0.1
pypdf==4.2.0
langchain-pinecone==0.1.1